Changed all files to handle new SemesterEnum

+++ src/main/java/nus/climods/model/module/UserModule.java
@@ -57,7 +57,7 @@ public class UserModule {
 
         listModule = optionalModule.get();
 
-        if (!listModule.getSemesters().contains(semester.getValue().intValue())) {
+        if (!listModule.getSemesters().contains(semester.getValue())) {
             throw new CommandException(MESSAGE_MODULE_NOT_OFFERED);
         }
